pyproject.toml
[tool.poetry]
name = "liturgical-calendar"
version = "0.1.0"
description = "Library to determine liturgical dates and colours for the Anglican Church of England"
authors = ["Jonathan Gazeley <me@jonathangazeley.com>"]
readme = "README.md"
packages = [{include = "liturgical_calendar"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
liturgical_calendar = 'liturgical_calendar.liturgical:main'
